fix: separation of logic and JSX
